Allow anonymous admin access until the first player connects in-game
Administrative access (i.e. start/stop/create/delete games, install modules, change MOTD and port) on a new server instance (no serverAdmins.json file) is public; as soon as a client connects through the regular game client (so an identity is generated), admin actions will be restricted to that player, which would need to authenticate to the web interface using the generated identity to be allowed to perform said actions.